server: convert user id to int

+++ server/auth.go
@@ -11,7 +11,7 @@ import (
 // checkAuth will return true if the user is authorized and false if not.
 // Authorization is determined by checking redis for the user id and token.
 // The user id is the key and the token is the value.
-func checkAuth(rdb *redis.Client, userID int64, token string) (bool, error) {
+func checkAuth(rdb *redis.Client, userID int, token string) (bool, error) {
 	result, err := rdb.Get(context.Background(), fmt.Sprintf("%d", userID)).Result()
 	if err == redis.Nil {
 		return false, nil
@@ -23,7 +23,7 @@ func checkAuth(rdb *redis.Client, userID int64, token string) (bool, error) {
 
 // createAuthToken will generate a random token and store it in Redis using the given id as the key.
 // Return value (token, err).
-func createAuthToken(id int64, rdb *redis.Client) (string, error) {
+func createAuthToken(id int, rdb *redis.Client) (string, error) {
 	token, err := generateToken(tokenBytes)
 	if err != nil {
 		return token, errors.New("generateToken error: " + err.Error())
